refactor: create plan::schedule submodule
